feat(pn-6975): refactor of api folder (#860)
